\name{additiveDEA-package}
\alias{additiveDEA-package}
\alias{additiveDEA}
\docType{package}
\title{
\packageTitle{additiveDEA}
}
\description{
\packageDescription{additiveDEA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{additiveDEA}
\packageIndices{additiveDEA}
The additiveDEA package provides functions for calculating efficiency with two types of additive Data Envelopment Analysis models: (i) Generalized Efficiency Measures: unweighted additive model (Cooper et al., 2007), Range Adjusted Measure (Cooper et al., 1999), Bounded Adjusted Measure (Cooper et al., 2011), Measure of Inefficiency Proportions (Cooper et al., 1999), and the Lovell-Pastor Measure (Cooper et al., 1999; Lovell and Pastor, 1995); and (ii) the Slacks-Based Measure (Tone, 2001). The functions provide several options: (i) constant and variable returns to scale; (ii) fixed (non-controllable) inputs and/or outputs; (iii) bounding the slacks so that unrealistically large slack values are avoided; and (iv) calculating the efficiency of specific Decision-Making Units (DMUs), rather than of the whole sample. Package additiveDEA also provides a function for reducing computation time when datasets are large
}
\author{
\packageAuthor{additiveDEA}

Maintainer: \packageMaintainer{additiveDEA}
}
